+++ rickled/cli/serve.py
@@ -25,7 +25,7 @@ def serve(args):
                 _input = sys.stdin.read()
             rick = Rickle(_input, load_lambda=args.LOAD_LAMBDA, RICKLE_PATH_SEP='/')
 
-        if args.b:
+        if args.BROWSER:
             import webbrowser
 
             scheme = 'https' if args.CERTIFICATE and args.PRIVATE_KEY else 'http'
@@ -34,7 +34,7 @@ def serve(args):
         serve_rickle_http(rickle=rick,
                           port=args.PORT,
                           interface=args.HOST,
-                          serialised=args.s,
+                          serialised=args.SERIALISED,
                           output_type=args.OUTPUT_TYPE,
                           path_to_certificate=args.CERTIFICATE,
                           path_to_private_key=args.PRIVATE_KEY,